This patch updates the `Role` model to use the `Webmozart\Assert\Assert` library for type-checking and validation. It also adds a new filter called "created_from" that allows filtering users by their creation date, and a new filter called "created_until" that allows filtering users by their creation date.
The patch also updates the `User` model to use the `Webmozart\Assert\Assert` library for type-checking and validation. It adds a new method called `getTable()` that returns the table name of the user model, and it also adds a new method called `casts()` that returns an array of casts for the user model.

The patch also updates the `Permission` model to use the `Webmozart\Assert\Assert` library for type-checking and validation. It adds a new method called `getTable()` that returns the table name of the permission model, and it also adds a new method called `casts()` that returns an array of casts for the permission model.

The patch also updates the `PermissionRole` model to use the `Webmozart\Assert\Assert` library for type-checking and validation. It adds a new method called `getTable()` that returns the table name of the permission role model, and it also adds a new method called `casts()` that returns an array of casts for the permission role model.

Finally, the patch updates the `User` model to use the `Webmozart\Assert\Assert` library for type-checking and validation. It adds a new method called `getTable()` that returns the table name of the user model, and it also adds a new method called `casts()` that returns an array of casts for the user model.